requirements.txt
asyncio
discord.py
yt-dlp
PyNaCl
python-dotenv
Flask